Move source files to src/, add flake8 config and update requirements
