+++ requirements.txt
@@ -1,2 +1,43 @@
+asciitree==0.3.3
+astmonkey==0.3.6
+astroid==2.12.12
+attrs==22.1.0
+black==22.10.0
+bytecode==0.13.0
+click==8.1.3
+commonmark==0.9.1
+exceptiongroup==1.0.1
+ffmpeg-python==0.2.0
 filetype==1.2.0
+flake8==5.0.4
+future==0.18.2
+iniconfig==1.1.1
+jellyfish==0.9.0
+Jinja2==3.1.2
+lazy-object-proxy==1.8.0
+MarkupSafe==2.1.1
+mccabe==0.7.0
+MutPy-Pynguin==0.7.1
+mypy-extensions==0.4.3
+networkx==2.8.8
+packaging==21.3
+pathspec==0.10.1
 Pillow==9.3.0
+platformdirs==2.5.3
+pluggy==1.0.0
+pycodestyle==2.9.1
+pydot==1.4.2
+pyflakes==2.5.0
+Pygments==2.13.0
+pynguin==0.27.0
+pyparsing==3.0.9
+pytest==7.2.0
+PyYAML==6.0
+rich==12.6.0
+simple-parsing==0.0.20
+termcolor==2.1.0
+tomli==2.0.1
+typing-inspect==0.8.0
+typing_extensions==4.4.0
+-e git+ssh://git@github.com/oliveox/watermarker-goes-boom.git@1397cdce4c8ca6f7239a06f264cbadb89f325d93#egg=watermarker
+wrapt==1.14.1
