+++ tox.ini
@@ -1,0 +1,4 @@
+[flake8]
+max-line-length = 120
+# T001 print found.
+extend-ignore = T001
